built the taskForGETMethod
taking reused code from get methods from loginviewcontroller, set a variable parameterswithAPIKey to contain a dictionary with the API key. build the request through NSMutableRequest, changing withPathExtension to method parameter that is being passed in. the request task is similar, changing the displayError function to a sendError function. using completionHandlerForGET method to send the error back to wherever this request is called. parsing the data is handled by convertDataWithCompletionHandler closure.